requirements.txt
sqlite3
sys
PyQt5.QtWidgets